Cleaner and tested azure plugin
